test(framework-delegate): migrate
